No user object causes error
Skip getting user session from hitl table if no user object on event.

When using modules like Webhook that does not attach a user object to the event it causes an error in incomingMiddlewares. Hitl tries to create new session but errors in knex call as user.id is undefined.